+++ src/main/kotlin/com/pestphp/pest/runner/PestConsoleProperties.kt
@@ -24,4 +24,9 @@ class PestConsoleProperties(
     override fun createRerunFailedTestsAction(consoleView: ConsoleView?): AbstractRerunFailedTestsAction? {
         return consoleView?.let { PestRerunFailedTestsAction(it, this) }
     }
+
+    override fun isPrintTestingStartedTime(): Boolean {
+        return false
+    }
+
 }